fixed bug with updating questions state

+++ components/Questions/QuestionItem.tsx
@@ -5,17 +5,29 @@ import { DeleteQuestion, GetQuestionsInOrgURL } from "../../adapters/Questions";
 import useStore from "../../utils/store";
 import { mutate } from "swr";
 import { DynamoNewQuestion } from "../../types/dynamo";
+import UpdateQuestionModal from "./UpdateQuestionModal";
+
 export default function QuestionItem({
   question,
 }: {
   question: DynamoNewQuestion;
 }) {
-  const openUpdateQuestionmodal = useStore(
-    (state) => state.openUpdateQuestionmodal
+  const setCurrentQuestion = useStore((state) => state.setCurrentQuestion);
+  const openUpdateQuestionModal = useStore(
+    (state) => state.openUpdateQuestionModal
   );
+
+  const handleEdit = () => {
+    setCurrentQuestion(question);
+    openUpdateQuestionModal();
+  };
+
   const [isHovering, setIsHovering] = useState(false);
 
   const handleDelete = async (questionId: string) => {
+    if (!confirm("Are you sure you want to delete this question?")) {
+      return;
+    }
     try {
       const data = await DeleteQuestion(questionId);
       alert(data.message);
@@ -24,17 +36,16 @@ export default function QuestionItem({
       alert(error.response.data.message);
     }
   };
+
   // Essentially fill in all the details of the modal and then open it
   return (
     <li
       onMouseEnter={() => setIsHovering(true)}
       onMouseLeave={() => setIsHovering(false)}
-      className="px-4 my-2 border-2 rounded-lg bg-white hover:bg-sky-50 transition ease-in-out duration-300 flex justify-between"
+      className="border rounded-lg shadow-sm px-4 py-2max-w-lg mx-auto my-2 bg-white hover:bg-sky-50 transition ease-in-out duration-300 flex justify-between"
     >
       <div className=" py-2  h-full relative focus-within:ring-2 focus-within:ring-blue-500">
         <h3 className="text-lg font-semibold text-dark">
-          {/* <a href="#" className="hover:underline focus:outline-none"> */}
-          {/* Extend touch target to entire panel */}
           <span className="absolute inset-0" aria-hidden="true" />
           {question?.GSI1SK}
         </h3>
@@ -44,26 +55,27 @@ export default function QuestionItem({
           </p>
         )}
         <p className="text-md text-red-300 line-clamp-2 mt-1">
-          {question?.questionId}
+          ID: {question?.questionId}
         </p>
       </div>
+      <div className="flex justify-center items-center ">
+        <button
+          onClick={handleEdit}
+          className="rounded-full hover:bg-white text-blue-500 transition ease-in-out duration-200 px-3 py-3 text-md"
+        >
+          <PencilAltIcon className="w-6 h-6" />
+        </button>
+        <button
+          onClick={() => handleDelete(question?.questionId)}
+          className="rounded-full hover:bg-white text-red-500 transition ease-in-out duration-200 px-3 py-3 text-md"
+        >
+          <TrashIcon className="w-6 h-6" />
+        </button>
+      </div>
 
-      {isHovering && (
-        <div className="flex justify-center items-center ">
-          <button
-            onClick={openUpdateQuestionmodal}
-            className="rounded-full hover:bg-white text-blue-500 transition ease-in-out duration-200 px-3 py-3 text-md"
-          >
-            <PencilAltIcon className="w-6 h-6" />
-          </button>
-          <button
-            onClick={() => handleDelete(question?.questionId)}
-            className="rounded-full hover:bg-white text-red-500 transition ease-in-out duration-200 px-3 py-3 text-md"
-          >
-            <TrashIcon className="w-6 h-6" />
-          </button>
-        </div>
-      )}
+      {/* {isHovering && (
+ TODO
+      )} */}
     </li>
   );
 }
